add float value for read with modbus

+++ src/main/java/com/modbus/reports/model/automationProtocols/modbus/Protocols.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.concurrent.ExecutionException;
 
 public interface Protocols {
-    List<Integer> transfer(String ip, TableView<Variables> tbVariables) throws InterruptedException, ExecutionException;
+    List<Float> transfer(String ip, TableView<Variables> tbVariables) throws InterruptedException, ExecutionException;
 
     boolean isConnect();
 }
