+++ src/main/java/com/modbus/reports/service/ProtocolsServiceImpl.java
@@ -22,7 +22,7 @@ public class ProtocolsServiceImpl implements ProtocolsService{
     public void setNode(String node) {
         this.node = node;
     }
-    public List<Integer> connecting () throws InterruptedException, ExecutionException {
+    public List<Float> connecting () throws InterruptedException, ExecutionException {
         return protocols.transfer(node, tbVariables);
     }
     @Override
